Add --message flag
Potentially useful for story IDs, to link commits to issue trackers.

This commit adds the argument to git-mob only, and even then not in
picker mode. I omitted git-solo because it didn't seem worth the effort,
it would just be indirection over `echo something >> ~/.gitmessage`. I
ommitted picker mode because it would have required some more
scaffolding and I was lazy. We can always add this later if desired.